+++ controlplane/monitor/internal/worker/worker.go
@@ -6,6 +6,8 @@ import (
 
 	devicetelemetry "github.com/malbeclabs/doublezero/controlplane/monitor/internal/device-telemetry"
 	internettelemetry "github.com/malbeclabs/doublezero/controlplane/monitor/internal/internet-telemetry"
+	"github.com/malbeclabs/doublezero/controlplane/monitor/internal/serviceability"
+	"github.com/prometheus/client_golang/prometheus"
 )
 
 type Watcher interface {
@@ -25,8 +27,20 @@ func New(cfg *Config) (*Worker, error) {
 		return nil, err
 	}
 
+	serviceabilityWatcher, err := serviceability.NewServiceabilityWatcher(&serviceability.Config{
+		Logger:         cfg.Logger,
+		Serviceability: cfg.Serviceability,
+		Interval:       cfg.Interval,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	deviceTelemetryMetrics := devicetelemetry.NewMetrics()
+	deviceTelemetryMetrics.Register(prometheus.DefaultRegisterer)
 	deviceTelemetryWatcher, err := devicetelemetry.NewDeviceTelemetryWatcher(&devicetelemetry.Config{
 		Logger:          cfg.Logger,
+		Metrics:         deviceTelemetryMetrics,
 		LedgerRPCClient: cfg.LedgerRPCClient,
 		Serviceability:  cfg.Serviceability,
 		Telemetry:       cfg.Telemetry,
@@ -36,8 +50,11 @@ func New(cfg *Config) (*Worker, error) {
 		return nil, err
 	}
 
+	internetTelemetryMetrics := internettelemetry.NewMetrics()
+	internetTelemetryMetrics.Register(prometheus.DefaultRegisterer)
 	internetTelemetryWatcher, err := internettelemetry.NewInternetTelemetryWatcher(&internettelemetry.Config{
 		Logger:                     cfg.Logger,
+		Metrics:                    internetTelemetryMetrics,
 		LedgerRPCClient:            cfg.LedgerRPCClient,
 		Serviceability:             cfg.Serviceability,
 		Telemetry:                  cfg.Telemetry,
@@ -49,6 +66,7 @@ func New(cfg *Config) (*Worker, error) {
 	}
 
 	watchers := []Watcher{
+		serviceabilityWatcher,
 		deviceTelemetryWatcher,
 		internetTelemetryWatcher,
 	}